split segmentation and base class
